Fix small rounding errors causing traveler to miss trips by fractions of a second
Due to small rounding errors bubbling up from core Network Analyst,
sometimes travelers miss transit connections by small fractions of a
second, particularly in cases where in reality they are remaining on the
same bus.  This fix is a brute-force solution to the problem.  When
considering what trips are available, the evaluator subtracts half a
second off the current time so as to simply avoid these tiny errors.