New accessibility matrix tool
Count the number of destinations reachable from each origin by transit
and walking. The tool calculates an Origin-Destination Cost Matrix for
each start time within a time window because the reachable destinations
change depending on the time of day because of the transit schedules.
The output gives the total number of destinations reachable at least
once as well as the number of destinations reachable at least 10%, 20%,
...90% of start times during the time window.  The number of reachable
destinations can be weighted based on a field, such as the number of
jobs available at each destination.  The tool also calculates the
percentage of total destinations reachable.