Fix handling of tomorrow and yesterday
In the CountTripsAtStops() function, it makes a list of all the trips
that are running on the day designated for the analysis, and then it
also grabs lists of trips for the previous day and the following day
because of the way service can wrap around (ie, trips that began in the
late hours of yesterday can still be running in the early hours of this
morning).

After getting those trip lists, it calls another function that runs a
SQL query to get all the stop times on those trips that are within the
time window.  I think I made a copy-paste error here a long time ago.
I’m feeding today’s trip list into the calls to this function for
yesterday and tomorrow.

This change feeds the correct trip lists into the correct calls to the
other function.