ValidateTableName was happening too soon
ValidateTableName removes characters from a string that are invalid for
fgdb feature class names.  Sometimes route_short_name has weird
characters in it, so we need to call this before determining the output
feature class names.  However, by calling it directly on the
route_short_name, sometimes it was adding the letter T to the
route_short_name when the route_short_name was a number, even though
when you prepend Stops_ and Buffers_ to a number, it's a valid name.
So, call ValidateTableName after adding the prefixes.