For unknown route_type values, use int instead of str
When the user's data has route_type values that don't match the ones listed in the GTFS spec, my code sets the route_type to 100 as just a random value to track them.  It was setting it to a string value, '100', instead of an integer, and this caused these routes to get ignored later when generating polylines.