Update linear referencing again
The Linear Referencing GP tool didn't do what I wanted, so intead, use the python geometry objects in a smarter way.  Primarily, measure the stop along the line, but then, retrieve the line segment up to that point and grab the geodesic length of it.  Don't rely on M values.