Bugfix x2 on CountTripsAtStopsByRouteAndDirection
(old) line 232 was trying to sort dict keys, but the keys are tuples, which don't have an obvious sort order. This crashed with `TypeError: '<' not supported between instances of 'str' and 'int'` as of Python 3.7. My solution was simply not to sort the dict keys: `for key in final_stop_freq_dict:` (`.keys()` is also unnecessary as iterating over a dict yields the keys). It would alternatively be possible to specify the optional `key` parameter of `sorted()` to establish an explicit basis for sorting the keys, but I don't think the key order actually matters. It affects the order of rows in the output feature class, but again, I don't think that order matters either.

The other fix has to do with GTFS feeds with a mix of blank and non-blank direction_ids. direction_id needs to be numeric (0 or 1) or null, but my test GTFS feed had some trips with direction_id = ''. This caused the script to fail farther down when it tried to store '' (an empty string) in a SHORT-type field. My solution was to take advantage of the preexisting check for whether direction_id is blank (circa line 115) and set direction_id = 0 if it is blank. There may be a safer/less destructive way to handle this...